AIBudgetTrackerBackend/requirements.txt
fastapi~=0.128.0
pydantic~=2.12.5
uvicorn~=0.40.0
gunicorn
google-genai
python-dotenv~=1.2.1
requests~=2.28.2
supabase~=2.27.2
argon2-cffi
passlib
python-dateutil~=2.8.2
pandas~=2.3.3
python-multipart==0.0.22
dotenv~=0.9.9
python-dateutil~=2.8.2
email-validator